NO MERGE: Debug path fixes
Inside of the docker container, codecov is run at the path
/workspace/tpm2-pytss. This seems to cause coverage uploads
not to resolve as the upload path is the CI's tpm2-pytss path.

I dropped a PWD before the codecov tool so we can see the path of
the repo there.

I also added a codecov --dump so we can get the contents of
coverage.xml.

Signed-off-by: William Roberts <william.c.roberts@intel.com>